% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_initDB.R
\name{.readSQLhistory}
\alias{.readSQLhistory}
\title{reads the sqlHistory table}
\usage{
.readSQLhistory(mydb)
}
\arguments{
\item{mydb}{database connection from DBI::dbConnect}
}
\description{
reads the sqlHistory table
}
