% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_initDB.R
\name{.updateSQLhistory}
\alias{.updateSQLhistory}
\title{updates the sqlHistory table}
\usage{
.updateSQLhistory(mydb, token, description)
}
\arguments{
\item{mydb}{database connection from DBI::dbConnect}

\item{token}{a number representing the time}

\item{description}{string with description of update}
}
\description{
updates the sqlHistory table
}
