% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_initDB.R
\name{.writeSQLdatabaseInit}
\alias{.writeSQLdatabaseInit}
\title{writes a SQLite database initialization}
\usage{
.writeSQLdatabaseInit(mydb, verbose = FALSE)
}
\arguments{
\item{mydb}{database connection from DBI::dbConnect}

\item{verbose}{logical to output extra information}
}
\description{
writes a SQLite database initialization
}
