% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_raw_root.R
\name{find_raw_root}
\alias{find_raw_root}
\title{Finds the directory with .rawdata2}
\usage{
find_raw_root(path = getwd())
}
\description{
Either it finds the .rawdata2 directory or the nearest .git directory
}
