% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_username.R
\name{git_username}
\alias{git_username}
\title{Get the configured Git username or system's username}
\usage{
git_username()
}
\value{
The configured Git username, the system username when Git has no
configured username, or \code{NA_character_} when the configured value is
empty.
}
\description{
Get the configured Git username or system's username
}
