% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_initDB.R
\name{raw.showHistoryDB}
\alias{raw.showHistoryDB}
\title{Prints the history of the SQLite database}
\usage{
raw.showHistoryDB(rawBase)
}
\arguments{
\item{rawBase}{rawBase class}
}
\description{
Prints the history of the SQLite database
}
