% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_export_legacy_rawData.R
\name{raw_export_legacy_rawData}
\alias{raw_export_legacy_rawData}
\title{Exports legacy rawData if available}
\usage{
raw_export_legacy_rawData(root_dir)
}
\description{
Exports legacy rawData if available
}
