% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_config.R
\name{raw_extensions_append}
\alias{raw_extensions_append}
\title{Add file extensions used by \code{raw_update()}}
\usage{
raw_extensions_append(extensions)
}
\arguments{
\item{extensions}{Character vector of extensions, with or without leading
dots.}
}
\value{
Invisibly, the complete configured extension vector.
}
\description{
Add file extensions used by \code{raw_update()}
}
