% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_getCRC.R
\name{raw_getCRC}
\alias{raw_getCRC}
\title{CRC code for file (per rawData)}
\usage{
raw_getCRC(filename)
}
\description{
Uses MD5 checksum or 32-byte MD5 hash for file, then shortens a bit
to store as a simple integer
}
