% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_find_ID.R
\name{raw_id_by_file}
\alias{raw_id_by_file}
\title{Retrieve RAW file IDs by filename or partial filename}
\usage{
raw_id_by_file(filename, ignore.case = FALSE, found_only = FALSE)
}
\description{
Retrieve RAW file IDs by filename or partial filename
}
