% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_init.R
\name{raw_init}
\alias{raw_init}
\title{Initializes the .rawdata2 directory}
\usage{
raw_init()
}
\value{
The normalized path to the initialized \code{.rawdata2} directory.
}
\description{
Initializes the .rawdata2 directory
}
