% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_path.R
\name{raw_path_trim}
\alias{raw_path_trim}
\title{Removes path that do not exist; use with care}
\usage{
raw_path_trim()
}
\value{
Invisibly, the character vector of paths read before nonexistent
paths were removed from the catalogue.
}
\description{
Removes path that do not exist; use with care
}
