% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_register.R
\name{raw_export_register}
\alias{raw_export_register}
\title{exports the RAWdata register to the target_path}
\usage{
raw_export_register(target_path)
}
\description{
exports the RAWdata register to the target_path
}
