% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_file_by_ID.R
\name{raw_file_by_id}
\alias{raw_file_by_id}
\title{Retrieve a RAW filename by ID}
\usage{
raw_file_by_id(ID)
}
\arguments{
\item{ID}{could be either numeric ID or base64 ID2}
}
\description{
Retrieve a RAW filename by ID
}
