% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_files_read.R
\name{raw_files_read}
\alias{raw_files_read}
\title{Read the RAW file catalogue}
\usage{
raw_files_read()
}
\value{
A data frame containing the RAW file catalogue. If the catalogue
does not exist, an empty data frame with \code{file} and \code{sha256} columns is
returned.
}
\description{
Read the RAW file catalogue
}
