% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_register.R
\name{raw_import_register}
\alias{raw_import_register}
\title{imports the RAWdata register}
\usage{
raw_import_register(recursive = FALSE)
}
\arguments{
\item{recursive}{will search directories recursively}
}
\description{
imports the RAWdata register
}
