% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_initDB.R
\name{raw_initDB}
\alias{raw_initDB}
\title{Initialize SQL database}
\usage{
raw_initDB(quiet = FALSE)
}
\arguments{
\item{quiet}{suppress messages}

\item{rawBase}{object, use create_rawBase()}
}
\description{
Create and initialize a database, if none exists. The SQLite
database can store images and large data files, which would
otherwise slow down the R package. The database contains the
name of the package, includes the version number and ends
with .sqlite. Before creating a new database, it checks whether
a database already exists; it might have an older version. If
so, then it updates (renames) the file to correspond to the
latest version. The rawBase object stores information to what
is stored in this separate database, in case it gets lost or
needs to be recreated.
}
