% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_list.R
\name{raw_list}
\alias{raw_list}
\title{Lists all available files}
\usage{
raw_list(nLEN = 40)
}
\arguments{
\item{nLEN}{Maximum number of trailing characters to show for each filename.}
}
\value{
A data frame containing the ID, ID2, final directory component, and
shortened filename for files currently marked as found.
}
\description{
Lists all available files
}
