% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_path.R
\name{raw_paths_file}
\alias{raw_paths_file}
\title{Data file with RAW paths}
\usage{
raw_paths_file()
}
\value{
The path to the \code{RAW_paths.csv} file. The \code{.rawdata2} directory is
created when necessary.
}
\description{
Data file with RAW paths
}
