% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_path.R
\name{raw_paths_read}
\alias{raw_paths_read}
\title{Reads the RAW paths}
\usage{
raw_paths_read()
}
\value{
A data frame of registered RAW paths and their metadata. If the
catalogue does not exist, an empty data frame is returned.
}
\description{
Reads the RAW paths
}
