% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_update.R
\name{raw_update}
\alias{raw_update}
\title{Update the RAW file catalogue with SHA-256 checksums}
\usage{
raw_update(paths = c())
}
\arguments{
\item{paths}{list of paths with data files}
}
\description{
Update the RAW file catalogue with SHA-256 checksums
}
